Fix for group authentication
When creating a new group an entry is added to the "auth_group" table
with the first row automatically created for the anonymous group

----
auth_group
id [PK] serial, name character varying(80)
1,anonymous
----

The geonode_authorize_layer function creates the group_ids (integer[])
by querying the "groups_groupmember"."group_id" column.

The issue is that the group_id column does not account for the anonymous
group. So when GeoServer is running the prepared statement. The result
is actually "groups_groupmember"."group_id" -1. A quick fix is just to
add 1 to each value in the "group_ids", since the anonymous group will
always be the first entry.